Meeting_Minutes/requirements.txt
fastapi==0.111.0
fsspec==2024.5.0
hjson==3.1.0
langchain==0.2.0
loguru==0.7.2
Markdown==3.6
pandas==2.2.2
pillow==10.3.0
pydantic>=2.7.4
python-docx==1.1.2
python-dotenv==1.0.0
python-pptx==0.6.23
requests==2.32.2
tabulate==0.9.0
tenacity
tqdm==4.66.4
trafilatura==1.10.0
uvicorn==0.29.0
webvtt-py==0.5.0
deepeval==1.4.4
llama-index==0.10.33
llama-index-llms-azure-openai==0.1.10
llama-index-extractors-entity==0.1.2
llama-index-embeddings-azure-openai==0.1.9
llama-index-embeddings-huggingface==0.2.0
llama-index-readers-papers==0.1.6
llama-index-readers-web==0.1.18
llama-index-vector-stores-azureaisearch==0.1.7
azure-search-documents
azure-ai-documentintelligence
